backend/pyproject.toml
[project]
name = "freelance-clienthub"
version = "0.1.0"
description = "Freelance ClientHub FastAPI backend"
authors = [{ name = "Sai" }]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "motor>=3.5.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = ["pytest", "httpx"]

[tool.uvicorn]
app = "app.main:app"
host = "0.0.0.0"
port = 8000
reload = true
